requirements.txt
datasets>=2.14.0
pandas>=1.5.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
opacus>=1.5.0 